% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-cor_plot.R
\name{corr_plot}
\alias{corr_plot}
\alias{corr_plot.formula}
\alias{corr_plot.data.frame}
\alias{corr_pairs}
\alias{corr_plot2}
\title{Correlation Plot}
\usage{
corr_plot(x, ...)

\method{corr_plot}{formula}(x, data, cex.labels = NULL, ...)

\method{corr_plot}{data.frame}(
  x,
  ...,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  main = "",
  sub,
  xlab,
  ylab
)

corr_pairs(
  data,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  ...
)

corr_plot2(
  ...,
  main = "",
  type = "pearson",
  sig.level = NULL,
  r.level = 0.1,
  mar = c(1, 1, 1, 1),
  include.order = FALSE,
  method = "color",
  col = RColorBrewer::brewer.pal(100, "RdBu")
)
}
\arguments{
\item{x}{formula oder data.frame}

\item{...}{an prepare_data2}

\item{data}{a data matrix}

\item{jitter}{Rauschen}

\item{smooth, lines}{Anpassungslienien}

\item{lines}{Regressinsgerade}

\item{pch}{Symbole  pch=20}

\item{digits, method}{correlation}

\item{cex.cor, resize}{Fixe groese mit cex.cor, resize abhaengig von r-Wert}

\item{method}{c("circle", "square", "ellipse", "number", "shade", "color", "pie"),}

\item{stars, resize, cex.cor}{correlation formatierung}

\item{hist}{Histogram TRUE/FLASE}

\item{col.bar, col.bar.border, col.line, col.smooth}{Farben}

\item{main}{titel}

\item{type}{c("full", "lower", "upper"),}

\item{sig.level}{signifikanz}

\item{col}{RColorBrewer::brewer.pal(100, 'RdBu')}

\item{smooth}{Gezeichnete Lineie}

\item{digits}{Nachkommastellen in plot}

\item{stars}{Sternchen}

\item{order}{c("original", "AOE", "FPC", "hclust", "alphabet"),}

\item{diag}{FALSE,}
}
\value{
nix
}
\description{
Function for making a correlation plot starting from a  formula and a data.frame
}
\examples{
#' 
# require(stp25plot)
# require(stp25tools)
require(tidyverse)
n <- 500
set.seed(1)

dat<- 
  data.frame(
    a = rnorm(n)) |>
  mutate(
    b = a + rnorm(n),
    c = b / 2 + rnorm(n),
    d = c / 3 + rnorm(n),
    e = 2 - a + rnorm(n),
    f = e / 5  + rnorm(n)
  )

# cor(dat)
# cor_matrix <- Hmisc::rcorr(dat)
corr_plot( ~ a + e + f+ b + c + d,
           dat,
           resize=FALSE)

corr_plot2(dat)

#par(mfrow= c(1,2))
corr_plot2(dat, main = "patient", sig.level = .2)
corr_plot2(dat, main = "patient", 
           r.level = .2, 
           type="spearman",
           order = TRUE
) 

}
