% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-forest_plot.R
\name{ggplot_table}
\alias{ggplot_table}
\title{ggplot_table}
\usage{
ggplot_table(
  gparam,
  main = "",
  gof = NULL,
  scale.log = FALSE,
  main.hjust = 0.5,
  left.hjust = -0.2,
  right.hjust = 0,
  space.right = 0.15,
  space.left = 0.02,
  space.shift.left = 0,
  space.shift.n = 0.22,
  space.shift.est = 0.4,
  refLabel = "reference",
  rangeb = range(gparam$conf.low, gparam$conf.high, na.rm = TRUE),
  digits = 2,
  cex = 0.7,
  pch = 15,
  fontsize = cex,
  point.size = 4,
  point.size.n = FALSE,
  col.text = 1,
  col.plot = "gray20",
  col.fill = NULL,
  lty.hline = 1
)
}
\arguments{
\item{gparam}{Parameter Tabele mit var, estimate, p.value usw}

\item{main}{Ueerschrift}

\item{gof}{Godnes of fit als Character}

\item{scale.log}{log-Transformation of x-scale TRUE/FALSE}

\item{right.hjust, left.hjust}{hjust erster und letzter Text}

\item{space.right, space.left}{space Plot}

\item{refLabel}{Bezeichnung der Referenzklasse}

\item{digits}{Digits}

\item{point.size, point.size.n}{point.size.n=TRUE Variable groesse der Punkte}

\item{col.text, col.plot, col.fill}{an ggplot c("#FFFFFF33", "#00000033") 
HEX8 Code die letzten stellen 33 sind fuer die Farb-Intensitaet}

\item{cpositions}{shift Position left side (Text: Item - levels - estimate)}
}
\value{
ggplot
}
\description{
In forest_plot( verwendet)
Gestohlen von survminer::ggforest()
}
\examples{

ggplot_table(
data.frame(
  var = c("Intercept", "Sex", "Sex", "Alter"),
  level = c(NA, "male", "female", NA),
  N = c(NA, 25, 47, 25+47),
  # pos = c(1, 1, 2, 1),
  estimate = c(-.77, NA, .51 , .4),
  conf.low = c(-1.53, NA, -.17, .2),
  conf.high = c(-0.1, NA, 1.2, .6),
  p.value = c(0.046, NA, 0.1407, 0.0021)
)
)

}
