% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-likertplot.R
\name{likert_col}
\alias{likert_col}
\title{likert_col}
\usage{
likert_col(n = 5, name = "RdBl", middle = mean(1:n), middle.color = "gray90")
}
\arguments{
\item{n}{Number of different colors in the palette}

\item{name}{A palette name from the lists below "RdBl"  ist   RdBl = c("Reds", "Blues")}

\item{middle, middle.color}{reference   "gray65"}
}
\value{
character
}
\description{
Farben:
  Greens,  Blues,  Reds,  Greys, Oranges, Purples
}
\examples{

par(mfrow=c(1,3))
barplot(cbind(1:5, rep(3,5)),  horiz = TRUE,  col=likert_col(5 , "RdBl"))
barplot(cbind(1:3, rep(3,3)),  horiz = TRUE,  col=likert_col(3 , "RdBl"))
barplot(cbind(1:8, rep(3,8)),  horiz = TRUE,  col=likert_col(8 , "RdBl"))



}
