% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-likertplot.R
\name{likertplot}
\alias{likertplot}
\alias{likert_plot}
\title{Barcharts for Likert}
\usage{
likertplot(
  x = Item ~ .,
  data = NULL,
  main = "",
  ylab = "",
  sub = "",
  xlab = if (horizontal) {
     if (as.percent) 
         "Prozent"
     else "Anzahl"
 }
    else "",
  ylim = NULL,
  xlim = NULL,
  col = NULL,
  rightAxis = FALSE,
  positive.order = NULL,
  include.order = NULL,
  decreasing = TRUE,
  as.percent = TRUE,
  auto.key = list(space = space, columns = columns, between = 1),
  ReferenceZero = NULL,
  reference.line.col = "gray65",
  col.strip.background = "gray97",
  wrap = TRUE,
  columns = 2,
  space = "top",
  horizontal = TRUE,
  between = list(x = 1 + (horizontal), y = 0.5 + 2 * (!horizontal)),
  par.settings = NULL,
  ...
)

likert_plot(
  ...,
  main = "",
  ylab = "",
  sub = "",
  xlab = if (as.percent) "Prozent" else "Anzahl",
  ylim = NULL,
  xlim = NULL,
  type = 1,
  col = NULL,
  rightAxis = FALSE,
  as.percent = TRUE,
  auto.key = list(space = space, columns = columns, between = 1),
  ReferenceZero = include.reference,
  reference.line.col = "gray65",
  col.strip.background = "gray97",
  wrap = TRUE,
  columns = 2,
  space = "top",
  horizontal = TRUE,
  positive.order = NULL,
  between = list(x = 1 + (horizontal), y = 0.5 + 2 * (!horizontal)),
  par.strip.text = list(lines = 1, cex = 0.8),
  par.settings = NULL,
  include.reference = NULL,
  include.total = FALSE,
  relevel = NULL,
  include.order = NULL,
  decreasing = TRUE,
  caption = "",
  include.table = FALSE,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.percent = TRUE,
  include.count = TRUE
)
}
\arguments{
\item{x}{formula}

\item{data}{daten}

\item{main, ylab, sub, xlab}{Beschriftung}

\item{col}{HH::brewer.pal.likert}

\item{rightAxis, as.percent, ReferenceZero, reference.line.col, col.strip.background}{an HH:::plot.likert.formula}

\item{positive.order}{das nicht verwenden!! - wird ueber Tbll_likert oder include.order = TRUE gesteuert.}

\item{include.order, decreasing}{Sortieren der Items}

\item{auto.key, columns, space}{columns = 2,}

\item{wrap}{Zeien Umbrechen
farbe("likert.blue.red", data$nlevels, middle = ReferenceZero)}

\item{...}{an stp25stat2::Likert}

\item{type}{nur in likert_plot Bei Gruppen Items als Zeilen => 1, oder Gruppen als Zeilen => 2}

\item{par.strip.text}{an  HH:::plot.likert.formula}

\item{include.total}{an stp25stat2:::Likert}

\item{relevel}{Uberschreibt die levels levels(x) <- relevel ist nur nur in likert_plot vorhanden}

\item{include.table, include.mean, include.n, include.percent, include.count, include.na, caption}{Tabelle ausgeben (output ist eine Altlast)}

\item{include.order}{sortiere muss die länge der Items entsprechen}
}
\value{
lattice Plot

HH likertplot (lattice-Plot)
}
\description{
Constructs and plots diverging stacked barcharts for Likert (copie from HH:::plot.likert.formula)
}
\details{
Die orginale Funktion hat bei der Sortierung (positive.order) einen Fehler.
}
\examples{

#require(stp25plot)
require(stp25stat2)
set.seed(1)
n <- 100
lvs <- c("--", "-", "o", "+", "++")
DF2 <- data.frame(
  Magazines = gl(length(lvs), 1, n, lvs),
  Comic.books = gl(length(lvs), 2, n, lvs),
  Fiction = gl(length(lvs), 3, n, lvs),
  Newspapers = gl(length(lvs), 5, n, lvs)
)



DF2$Comic.books[sample.int(n / 2)] <- lvs[length(lvs)]
DF2$Newspapers[sample.int(n / 2)] <- lvs[1]
DF2$Magazines[sample.int(n / 2)] <- lvs[2]

DF2 <- transform(DF2, Geschlecht = cut(rnorm(n), 2, Hmisc::Cs(m, f)))
Res1 <- Tbll_likert( ~ ., DF2[, -5])
Res2 <- Tbll_likert(. ~ Geschlecht, DF2)

# require(HH)  # ?likertplot
# class(Res2)
# windows(7, 3)
# attr(Res2, "plot")$results

likertplot(Item ~ . | Geschlecht,
           data = Res2,
            between=list(x=0))

# col = likert_col(attr(data, "plot")$nlevels, middle = ReferenceZero)

DF2 \%>\% likert_plot(Magazines, Comic.books, Fiction, Newspapers,
                    relevel = letters[1:5],
                    ReferenceZero = 1.5,
                    columns=5)

# DF2 \%>\%
#   Likert(Magazines, Comic.books, Fiction, Newspapers) \%>\%
#   likertplot()

sim_factor <-
function(n = 10,
         labels = c("--", "-", "o", "+", "++"),
         levels = seq_along(labels)) {
  factor(sample(levels, n, replace = TRUE), levels, labels)
}


DF <- data.frame(
  Employment.sector =
    sim_factor(
      n,
      c(
        "Academic (nonstudent)",
        "Business and industry",
        "Federal, state, and local government",
        "Private consultant/self-employed",
        "Other (including retired, students, not employed, etc.)")),
  Race = sim_factor(n, c("White","Asian","Black or African American","Other")),
  Education = sim_factor(n, c("Associate's and Bachelor's","Master's and Above")),
  Gender = sim_factor(n, c("Male","Female")),
  Prof.Recog = sim_factor(n, c("Not Important", "Important")),
  Question = sim_factor(n,
                        c(
                          "Strongly Disagree",
                          "Disagree" ,
                          "No Opinion",
                          "Agree Strongly", "Agree"
                        ))

)
All <- DF |> Summarise(Question, fun =table)
Employment.sector <- DF |> Summarise(Question~ Employment.sector, fun =table)
Race <- DF |> Summarise(Question~ Race, fun =table)
Gender <- DF |> Summarise(Question~ Gender, fun =table)
Education <- DF |> Summarise(Question~ Education, fun =table)
Prof.Recog <- DF |> Summarise(Question~ Prof.Recog, fun =table)



DF2 <-
  Rbind2(All, Employment.sector, Race, Gender, Education, Prof.Recog,
         .id = "Subtable") |>
  mutate(
    Question = dplyr::coalesce(Employment.sector, Race, Gender, Education, Prof.Recog ),
    Question = as.character(Question),
    Subtable = as.character(Subtable)
  ) |>
  select(Subtable, Question, `Strongly Disagree`, Disagree, `No Opinion`, `Agree Strongly`, Agree)


DF2$Question[1] <- " All Survey Responses"
# DF2$Question = factor(DF2$Question)
# DF2$Subtable = factor(DF2$Subtable)
stp25plot::likertplot(Question ~ . | Subtable, DF2,
                      scales=list(y=list(relation="free")), layout=c(1,6),
                      #  positive.order=TRUE,
                      between=list(y=0),
                      strip=FALSE, strip.left=strip.custom(bg="gray97"),
                      par.strip.text=list(cex=.6, lines=5),
                      main="Is your job professionally challenging?",
                      ylab=NULL,
                      wrap =FALSE,
                      sub="This looks better in a 10inx7in window")

DF2 \%>\% likert_plot(Magazines, Comic.books, Fiction, Newspapers)

}
