% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-auto-plot.R
\name{marginal_plot}
\alias{marginal_plot}
\title{lattice marginal plot}
\usage{
marginal_plot(
  ...,
  par.settings = bw_theme(farbe()),
  auto.key = list(lines = TRUE),
  plot.points = FALSE,
  ref = TRUE,
  cut = 0,
  origin = 0,
  xlab = NULL,
  ylab = NULL,
  type = NULL,
  main = "marginal distributions",
  subset = TRUE,
  as.table = TRUE,
  subscripts = TRUE,
  default.scales = list(relation = "free", abbreviate = TRUE, minlength = 5, rot = 0, cex
    = 0.75, tick.number = 3),
  layout = NULL,
  lattice.options = list(layout.heights = list(axis.xlab.padding = list(x = 0),
    xlab.key.padding = list(x = 0)))
)
}
\arguments{
\item{...}{an prepare_data2}

\item{par.settings, auto.key}{an lattice sefault ist
par.settings = stp25output::set_lattice()}
}
\value{
lattice Plot
}
\description{
lattice marginal plot
}
\examples{

enviro <- environmental
## make an ordered factor (so it will not be reordered)
enviro$smell <- cut(enviro$ozone, breaks = c(0, 30, 50, Inf),
                    labels = c("ok", "hmmm", "yuck"), ordered = TRUE)
enviro$is.windy <- factor(enviro$wind > 10,
                          levels = c(TRUE, FALSE), labels = c("windy", "calm"))
head(enviro)

# marginal.plot(enviro[,1:5], data = enviro, groups = is.windy,
#               auto.key = list(lines = TRUE))
marginal_plot(enviro, ozone, radiation, is.windy, wind, smell, by=~temperature)
}
