% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-multi-barchart.R
\name{multi_barplot}
\alias{multi_barplot}
\title{multi_barplot}
\usage{
multi_barplot(
  ...,
  reorder = TRUE,
  last = NULL,
  main = NULL,
  ylab = "",
  include.percent = FALSE,
  origin = 0,
  xlab = if (include.percent) "Percent" else "Count",
  xlim = NULL,
  ylim = NULL,
  wrap = TRUE,
  use.level = 1
)
}
\arguments{
\item{...}{an Summarise}

\item{reorder, last}{an reorder2}

\item{main, ylab, origin, xlab}{an Lattice}

\item{include.percent}{summary}
}
\value{
lattice plot
}
\description{
multi_barplot
}
\examples{

#'set.seed(1)
n<-99
DF<- data.frame(
  Magazines = rbinom(n, 1,prob=.75),
  Comic.books =rbinom(n, 1,prob=.25),
  Fiction = rbinom(n, 1,prob=.5),
  Sonstiges = rbinom(n, 1,prob=.35)
)


DF <- transform(DF, sex = cut(rnorm(n), 2, c("m", "f")))


multi_barplot(  DF, .~ sex, last="Sonstiges")


}
