% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-auto-plot.R
\name{multi_uv_plot}
\alias{multi_uv_plot}
\title{Hilfsfunktion}
\usage{
multi_uv_plot(
  data,
  measure.vars,
  group.vars,
  row_name,
  col_name,
  group.class,
  measure,
  origin,
  xlab,
  ylab,
  type,
  default.scales,
  lattice.options,
  par.settings,
  include.n,
  cex.main,
  layout,
  par.strip.text,
  include.percent,
  levels.logical,
  labels.logical
)
}
\description{
Im wesentlichen ist das eine Kopie von oben nur die Formeln sind vertauscht 
und die auswahl an verschiede Plots ist nicht möglich.
 @noRd
}
