% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart.text.R
\name{panel.barchart.text}
\alias{panel.barchart.text}
\title{panel.barchart.text}
\usage{
panel.barchart.text(
  x,
  y,
  box.ratio = 1,
  horizontal = TRUE,
  stack = FALSE,
  groups = NULL,
  cex = 0.7,
  col = 1,
  digits = 2,
  fmt = paste0("\%.", digits, "f"),
  labels = NULL,
  prefix = NULL,
  suffix = NULL,
  offset.col = c("black", "white"),
  offset.tresh = NULL,
  offset = NULL,
  ...
)
}
\arguments{
\item{x, y, box.ratio, horizontal, stack, groups}{kommt alles von lattice}

\item{cex, col}{an panel.text die Groesse und die Farbe der Schrift}

\item{digits, fmt}{digits geht an fmt siehe  sprintf}

\item{labels}{Sortierung der Variablen beachten}

\item{prefix, suffix}{Text dekoration}

\item{offset}{Text inerhalb oder ausserhalb der Balken default ist ofoset= 0.5,}

\item{...}{an panel.text  lineheight, font, fontfamily, fontface, adj}
}
\value{
lattice
}
\description{
Zahlen in Barplots
}
\examples{

# graphics.off()

#--http://www.ats.ucla.edu/stat/r/faq/barplotplus.htm
#-- 24.06.2015 08:27:01

# library(stp25plot)
# require(lattice)
set.seed(2)

DF <-
  data.frame(
    value = runif(2 * 5, min = 20, max = 80),
    sex = factor(rep(c("male", "female"), times = 5)),
    variable = factor(rep(
      c(
        n = "Neurotizismus",
        e = "Extraversion",
        o = "Offenheit",
        g = "Gewissenhaftigkeit",
        a = "Vertraeglichkeit"
      ),
      times = 2
    ))
  )




my_key <- list(
  columns = 2,
  space = "top",
  cex = 0.8,
  size = 1.4,
  adj = 1,
  between = 0.2,
  between.colums = 0.1
)



#' windows(5, 6)
#' barchart(
#'   reorder(variable, value) ~ value,
#'   subset(DF,
#'          sex == "male"),
#'
#'   box.ratio = 2,
#'   xlim = c(-5, 100),
#'   origin = 0,
#'
#'   #' par.settings=colorset,
#'   panel = function(...)   {
#'     panel.barchart(...)
#'     panel.barchart.text(..., digits = 1, suffix = " \%")
#'   }
#' )
windows(5, 6)
barchart(
  variable ~ value,
  DF,
  groups = sex,
  box.ratio = 2,
  xlim = c(-5, 100),
  origin = 0,
  auto.key = my_key,
  #' par.settings=colorset,
  panel = function(...)   {
    panel.barchart(...)
    panel.barchart.text(...,
                        digits = 1,
                        suffix = " \%",
                        offset = -4)
  }
)




windows(5, 6)
barchart(
  value  ~ variable,
  DF,
  groups = sex,
  stack = TRUE,
  box.ratio = 2,
  ylim = c(-5, 100),
  origin = 0,
  auto.key = my_key,
  #' par.settings=colorset,
  panel = function(...)   {
    panel.barchart(...)
    panel.barchart.text(..., digits = 1, suffix = " \%")
  }
)





dat<- stp25tools::get_data("
id  value  variable  Freq
1             'weniger als heute'  Telefonauskunft  68.333333
2  'im gleichen Ausmaß wie heute'  Telefonauskunft  25.000000
3                'mehr als heute'  Telefonauskunft   0.000000
4                  'es gab keine'  Telefonauskunft   6.666667
5             'weniger als heute'  Elektronische  28.333333
6  'im gleichen Ausmaß wie heute'  Elektronische  53.333333
7                'mehr als heute'  Elektronische   3.333333
8                  'es gab keine'  Elektronische  15.000000
9             'weniger als heute'  Videokonferenzen   0.000000
10 'im gleichen Ausmaß wie heute'  Videokonferenzen   1.000000
11               'mehr als heute'  Videokonferenzen   0.000000
12                 'es gab keine'  Videokonferenzen 100.000000
13            'weniger als heute'  Kurzmessenger   5.000000
14 'im gleichen Ausmaß wie heute'  Kurzmessenger   3.333333
15               'mehr als heute'  Kurzmessenger   0.000000
16                 'es gab keine'  Kurzmessenger  91.666667")

barchart(value~Freq|variable, dat, 
         par.strip.text = list(lines = 2.5, cex=.75 ),
         xlab ="Prozent", origin =0,
         panel = function(...)   {
           panel.barchart(...)
           panel.barchart.text(..., digits = 0, suffix = " \%", 
                               offset = c(-2.5, .4),
                               offset.tresh=20)
         }
         
)

}
