% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart.value.R
\name{panel.barchart.value}
\alias{panel.barchart.value}
\title{panel.barchart.value}
\usage{
panel.barchart.value(
  ...,
  cex = 0.7,
  col = 1,
  adj = c(0.5, 0.5),
  pos = NULL,
  offset = 0.5,
  digits = "\%.02f",
  labels = NULL,
  prefix = NULL,
  suffix = NULL,
  fun = NULL
)
}
\arguments{
\item{...}{alles von lattice}

\item{cex}{die groesse}

\item{col}{col=1 Farbe der Schrift}

\item{adj}{adj= c(0.5, 0.5),}

\item{pos}{= NULL  pos = 4 ist hozizoltal rechts vom Balken, pos = 3 ist horizontal oben vom Balken}

\item{offset}{default ist ofoset= 0.5,}

\item{digits}{siehe auch \link{sprintf}}

\item{labels}{Sortierung der Variablen beachten}

\item{prefix, suffix}{Formatierung}

\item{fun}{Funktion zum Formatiren}
}
\value{
lattice
}
\description{
Zahlen in Barplots
}
\examples{
#graphics.off()
 
  #--http://www.ats.ucla.edu/stat/r/faq/barplotplus.htm
  #-- 24.06.2015 08:27:01

library(stp5)
library(plyr)
Start()
set.seed(2)
n<-20*3*2
DF<- upData(data.frame(n=runif(n, min = 1, max = 5),
e=runif(n, min = 1, max = 5),
o=runif(n, min = 1, max = 5),
g=runif(n, min = 1, max = 5),
a=runif(n, min = 1, max = 5),
treatment=gl(3, n/3, labels = c("UG1", "UG2", "KG"))[sample.int(n)],
sex=gl(2, n/2, labels = c("male", "female")
)
), labels=c(  n="Neurotizismus", 
             e="Extraversion",
             o="Offenheit fuer Erfahrungen",
             g= "Gewissenhaftigkeit",
             a="Vertraeglichkeit") )
DF[1,1:3] <-5;DF[1,4:5] <-1; DF[2:10,1] <-4.5
DF[n,4:5] <-5;DF[n,1:5] <-1
APA2(~., DF)

  # ( x <-Recast2(.~sex, DF, fun=mean,formula = sex ~variable ))
  #  xMatrix<-as.matrix(x[,-1])
  #windows(6,4)
  #par(las=2, mar=c(5,15,5,5))
  #
  #bp <- barplot2(xMatrix,
  #beside = TRUE, horiz = TRUE,
  #names.arg = colnames(x)[-1],
  ## plot.ci = TRUE, ci.u = upper, ci.l = lower,
  #col = c("lightblue", "mistyrose" ),
  #xlim = c(0, 5),
  #legend = x$sex
  #)
  #text(xMatrix,bp,round(xMatrix, 1),cex=1,pos=4) #' label on the bars themselves
  #title(main = "Mean math scores by SES and gender", font.main = 4)


( ag.df <-Recast2(n+e+o+g+a~sex, DF, fun=mean  ))

  #-------------------------------------------------------------------------------
  #g2<-ggplot(ag.df, aes(x = reorder(variable, value), y = value, fill=sex)) +
  #    geom_bar(stat = "identity", position=position_dodge()) +  #',width= .91
  #   scale_fill_manual(values=c("lightblue", "mistyrose" )) +
  #    #'scale_fill_discrete( name ="",
  #    #'                        labels=levels(ag.df$sex)) +
  #
  #    #' scale_fill_brewer(palette="Set1")+
  #    geom_text(aes(y=value, ymax=value-1, label=   sprintf("\%.02f", value) #'   # round(value,2)
  #                 ),
  #               position= position_dodge(width=0.9),
  #               vjust= 0,hjust=1, color="black" ,
  #               cex=3) +
  #    scale_y_continuous("" , limits=c(0, 4), breaks=0:4,
  #              labels=c("anfang", "2","3","4", "ende"))  +
  #    scale_x_discrete("" ) +
  #    theme(axis.text.y=element_text(colour="black"))+
  #    coord_flip()
  #windows(6,4)
  #g2

  #-------------------------------------------------------------------------------
barchart(value~reorder(variable, value),  ag.df,  groups=sex, box.ratio=2,
                          ylim=c(-0.9,4.1), origin=1,
             #' par.settings=colorset,
              panel=function(...)   {
                            panel.barchart(...)
                            panel.barchart.value(..., digist=2)
              }
               ,
              auto.key=list(columns=2, space="top",
                            cex=0.8, size=1.4, adj=1,
                            between=0.2, between.colums=0.1)
                            )



( ag.df <-Recast2(n+e+o+g+a~sex+treatment, DF, fun=mean  ))


windows(12,4)
barchart(reorder(variable, value)~value| treatment ,  ag.df,  groups=sex,  box.ratio=5,
                          xlim=c(0.9,4.1), origin=1,
             #' par.settings=colorset,
              panel= function(...) {
                panel.barchart(...)
                panel.barchart.value(..., digits=4)

             },
              auto.key=list(columns=2, space="top",
                            cex=0.8, size=1.4, adj=1,
                            between=0.2, between.colums=0.1)
                            )

# #'windows(4,4)
# x <- xtabs( ~Unfall.Gruppe, DF )
# x <-   data.frame((x))
# 
# barchart( Freq~Unfall.Gruppe, x,
#           origin=0,
#           panel=function(...)   {
#               panel.barchart(...)
#               panel.barchart.value(..., digits=2, pos= 3, suffix="\%", 
#                                    fun=function(x){
#                                        prz<- round(x/ sum(x)*100)
#                                        paste0(prz, "\% (",x, ")" )
#                                    }
#               )
#           }
#           
#           
# )
}
