% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.conf.int.R
\name{panel.conf.int}
\alias{panel.conf.int}
\alias{panel.median}
\title{panel.conf.int}
\usage{
panel.conf.int(
  x,
  y,
  ...,
  pch = 19,
  lwd = 2,
  conf.level = 0.95,
  length = 0.05,
  method = "mean"
)

panel.median(x, y, ..., pch = 19, conf.level = 0.95, length = 0.05)
}
\arguments{
\item{pch}{= 16,}

\item{conf.level}{= 0.95,}

\item{length}{= 0.05,}

\item{method}{="mean"}
}
\description{
Erstellt Mittelwertdiagramm mit Confidenzintervall
}
\examples{
library(stp5)

 n<-100
 set.seed(3)
 DF <- make.groups( P.F.ok=data.frame( value=rnorm(n,-.4),type=gl(2, n/2, labels = c("Control", "Treat")))
                    ,P.F.uk =data.frame( value=rnorm(n,.4),type=gl(2, n/2, labels = c("Control", "Treat")))
                    , P.M.ok=data.frame( value=rnorm(n),type=gl(2, n/2, labels = c("Control", "Treat")))
                    , P.M.uk=data.frame( value=rnorm(n,.2,.5),type=gl(2, n/2, labels = c("Control", "Treat"))))
 windows(8,3)
 bwplot(which~value|type  , DF, xlim=c(-1, 1), xlab= "Difference (Absolute Value) [mm]",
       panel=function(x,y, ...){
         # panel.bwplot(x, y, ...)
         panel.abline(v=0, lty=2)
          panel.conf.int(x,y,...)

       })



}
