% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.mean.R
\name{panel.mean}
\alias{panel.mean}
\title{panel.mean}
\usage{
panel.mean(
  x,
  y,
  fun = mean,
  horizontal = TRUE,
  lwd = reference.line$lwd,
  lty = reference.line$lty,
  col = trellis.par.get("superpose.symbol")$col,
  col.line = reference.line$col,
  type = "p",
  pch = trellis.par.get("superpose.symbol")$pch,
  ...,
  identifier = "linejoin"
)
}
\description{
Erstellt Mittelwertdiagramm Mittelwerte in bwplots copie von panel.average()
}
\examples{

 set.seed(1)
res = data.frame(coef=rnorm(99) + (-1):1,
                 habitat=sample(letters[1:4], 99, TRUE),
                 grp=c("W", "X", "Y"))





bwplot(coef ~ habitat | grp, data = res, type="p",
       panel=function(...){
         #  panel.bwplot(..., pch='|')
       
         panel.mean(...)
       }
)


}
