% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_pie_plot.R
\name{pie_plot}
\alias{pie_plot}
\alias{panel.piechart}
\alias{piechart}
\title{pie_plot}
\usage{
panel.piechart(x, y, labels = as.character(y), edges = 200,
  radius = 0.8, clockwise = FALSE, init.angle = if (clockwise) 90
  else 0, density = NULL, angle = 45, col = superpose.polygon$col,
  border = superpose.polygon$border, mar = c(2, 2, 2, 2) - 0.2,
  digits = 0, lty = superpose.polygon$lty, ...)

piechart(x, data = NULL, panel = "panel.piechart", ...)
}
\arguments{
\item{data}{Data.frame}

\item{formula}{Objekt}

\item{main}{Ueberschrift}
}
\description{
Standardgrafiken
}
\examples{
set.seed(2)
n<-20*3*2
DF<- upData(data.frame(n=runif(n, min = 1, max = 5),
                       e=runif(n, min = 1, max = 5),
                      o=runif(n, min = 1, max = 5),
                       g=runif(n, min = 1, max = 5),
                       a=runif(n, min = 1, max = 5),
                       treatment=gl(3, n/3, labels = c("UG1", "UG2", "KG"))[sample.int(n)],
                       sex=gl(2, n/2, labels = c("male", "female")
                       )
), labels=c(  n="Neurotizismus",
              e="Extraversion",
              o="Offenheit fuer Erfahrungen",
              g= "Gewissenhaftigkeit",
              a="Vertraeglichkeit") )
DF[1,1:3] <-5;DF[1,4:5] <-1; DF[2:10,1] <-4.5
DF[n,4:5] <-5;DF[n,1:5] <-1
APA2(~treatment, DF)
windows(8,8)
pie_plot(~treatment, DF)
windows(8,8)
pie_plot(~treatment, DF)
windows(8,8)
Balken(~treatment, DF)
windows(8,8)
Balken(~treatment, DF, type=2)
 #
windows(8,8)
pie_plot(~treatment, DF, init.angle=45, main="pie_plot")
# windows(8,8)
# pie_plot(~treatment+sex, DF, init.angle=45)
#
# windows(8,8)
# pie_plot(~treatment|sex+BI, DF, init.angle=45)
}
