% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.efflist.R
\name{plot.efflist}
\alias{plot.efflist}
\title{Modifiziert effect-Plot}
\usage{
\method{plot}{efflist}(x, selection, rows, cols, ask = FALSE,
  graphics = TRUE, lattice, xlab, ...)
}
\arguments{
\item{x}{Model fit}

\item{selection, rows, cols, ask, graphics}{alles nicht ändern}

\item{xlab}{xlabs Ändern}

\item{...}{}
}
\value{
Plot
}
\description{
Soll nur im Notfall geladen werden geht nicht Mehr
}
\examples{
library(effects)
library(gridExtra)
A = rnorm(100)
B = rnorm(100)
C = factor(rep(c("This", "That"), 50))

model = lm(A ~ B * C)

p1 <- plot(ef <- effect("B:C", model), x.var = "C")

ef$variables$C$levels <- c("foo", "bar")
levels(ef$x$C) <- c("foo", "bar")
p2 <- plot(ef, x.var = "C")

grid.arrange(p1, p2)


#ef$variables
#
ef[[1]]$variables$C$levels <- c("foo", "bar")
levels(ef[[1]]$x$C) <- c("foo", "bar")
windows(8, 8)
plot(ef, x.var = "C")


#-- Modifiziert wegen xlab

plot.efflist <- stp25:::plot.efflist

ef <- allEffects(lm(A ~ B + C))
plot(ef, xlab = c("Foo", "Bar"))
}
