% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC2.R
\name{plotROC2}
\alias{plotROC2}
\title{plotROC}
\usage{
plotROC2(
  x,
  ...,
  ylab = "Sensitivity (TPR)",
  xlab = if (legacy.axes) "1-Specificity (FPR)" else "Specificity (FPR)",
  main = NULL,
  lwd = 1.1,
  size = lwd,
  col = NULL,
  colour = if (is.null(col)) "gray50" else col,
  include.facet_wrap = FALSE,
  legend.position = "bottom",
  include.labels = TRUE,
  include.table = FALSE,
  include.order = TRUE,
  caption = "Optimal cut-point: Area Under the Curve (AUC), Youden Index (YI).",
  legacy.axes = TRUE,
  ann.segment = if (legacy.axes) c(0, 1, 0, 1) else c(0, 1, 1, 0)
)
}
\arguments{
\item{x}{a roc object from the roc function, or a list of roc objects.}

\item{...}{description}

\item{ylab, xlab, main}{character.   false positive fraction = FPF}

\item{lwd, size}{numeric.  lwd}

\item{col, colour}{character. Farbe colour ist für singel}

\item{include.facet_wrap}{logical. facet}

\item{legend.position}{character. position}

\item{include.labels}{logical.mit Label()erstellte Beschriftungen TRUE/FALSE}

\item{include.table, caption}{logical. Tabelle Ausgeben}

\item{include.order}{logical. sortieren nach auc
true positive fraction = TPF}

\item{legacy.axes, ann.segment}{logical.  x-achse 0 bis 1 oder 1 bis 0,  ann.segment = annotate("segment")}
}
\value{
ggplot
}
\description{
This function plots a ROC curve with ggplot2.
}
\details{
Stolen from ggroc.roc {pROC}

Calculate the empirical Receiver Operating Characteristic curve

Given a binary outcome d and continuous measurement m, computes the empirical
ROC curve for assessing the classification accuracy of m
 


Some R Packages for ROC Curves: https://rviews.rstudio.com/2019/03/01/some-r-packages-for-roc-curves/
}
\examples{

require(pROC)

 data(aSAH)
 roc(outcome ~ s100b, data=aSAH) |>
 plotROC2()
 
 # mehr Beispiele unter ?Tbll_roc
}
