% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-fliegen-schiss-plot.R
\name{plotSigBars}
\alias{plotSigBars}
\alias{panel.sig.bars}
\title{Der Fliegen-Schiss-Plot}
\usage{
plotSigBars(
  x,
  include.stars = TRUE,
  shift = TRUE,
  cex.stars = 0.9,
  offset = 0.2,
  ...
)

panel.sig.bars(
  x,
  include.stars = TRUE,
  shift = TRUE,
  cex.stars = 0.9,
  offset = 0.4,
  ...
)
}
\arguments{
\item{x}{Objekt lm oder emmeans}

\item{shift}{oben unden oder links rechts}

\item{cex.stars}{groesse}

\item{...}{nicht benutzt}

\item{stars}{Sternchen oder P-Weret}
}
\value{
plot-text
}
\description{
Der Fliegen-Schiss-Plot
}
