% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-fliegen-schiss-plot.R
\name{prism.plots}
\alias{prism.plots}
\title{Plot prism-like Plots}
\usage{
prism.plots(
  formula,
  data,
  fun = median,
  interquartile = TRUE,
  spreadfunc = function(x) {
     return(1.96 * sd(x)/sqrt(length(x)))
 },
  def.axis = TRUE,
  jitter = TRUE,
  add = FALSE,
  col = "gray",
  pch = 16,
  ...
)
}
\arguments{
\item{formula}{a formula object with the quantitative variable as the response variable (e.g., Var~group).}

\item{data}{a dataset containing the variables indicated in \code{formula}}

\item{fun}{what function should be used to indicate the center of the distribution. Defaults to median.}

\item{interquartile}{Should the interquartile range be plotted? Defaults to TRUE.}

\item{spreadfunc}{what function should be used to calculate the spread of the distribution? If interquartile=TRUE,
this argument will be ignored. The default (when not ignored) is to produce a 95\% confidence interval (1.96*sd(x)/sqrt(n)).}

\item{def.axis}{Logical. Should the default axes be used?}

\item{jitter}{Logical. Should the y values be jittered as well?}

\item{add}{Should the plot be added to an existing plot?}

\item{col}{What color should the dots be painted? Defaults to gray.}

\item{...}{other arguments passed to plot}
}
\description{
gestohlen von Dustin Fife
}
\examples{

require(lattice)
require(latticeExtra)
require(effects)

prism.plots(count ~ spray, data = InsectSprays, fun = mean)

Insect <- lm(count ~ spray, data = InsectSprays)
stripplot(
  count ~ spray,
  data = InsectSprays,
  # ylim=c(-10,70),
  ylab = "Insect count",
  xlab = "The type of spray",
  jitter.data = TRUE,
  panel = function(x, y, ...) {
    # panel.conf.int(x, y, ...)
    panel.stripplot(x, y, pch = 19, col = "gray50", ...)
    # panel.points(x,y, pch=19,...)
    # panel.mean(x,y,  ...)
    panel.median(x, y, ...)
    panel.sig.bars(Insect, include.stars = FALSE, offset = .4)
  }
)


p <- stripplot(
  count ~ spray,
  data = InsectSprays,
  jitter.data = TRUE,
  pch = 20,
  col = "gray50"
)
p + layer(panel.sig.bars(Insect, include.stars = FALSE, offset = .4))

p2 <- plot(effect("spray", Insect), lty = 0 , main = "Effectiveness of Insect Sprays")
p2 +  layer(panel.sig.bars(Insect, include.stars = FALSE, offset = .4))
}
