% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_profile_plot.R
\name{profile.plot}
\alias{profile.plot}
\title{Create a profile plot.}
\usage{
\method{profile}{plot}(x, text.left = NULL, text.right = NULL,
  highlight.col = -1, text.width.adjust = 35, colors = NULL,
  draw.segments = TRUE, draw.mid = TRUE, plot.legend = TRUE,
  legend.n.col = 4, sep = " - ", ...)
}
\arguments{
\item{x}{The matrix of data to plot. Each profile item is in a column with an appropriate column name.The column names will be used as the names for the profiles in the legend.}

\item{text.left}{text.left=NULL The text to appear on the left hand side (same length as nrow(x))}

\item{text.right}{text.right=NULL The text to appear on the right hand side (same length as nrow(x))}

\item{highlight.col}{highlight.col=-1 The column index of the entry to highlight. (Default, none highlighted)}

\item{text.width.adjust}{text.width.adjust=35 The adjustment factor for the spacing for the text on the left and right of the plot. If you  need more space, make this larger. If you need less, make it smaller.}

\item{colors}{colors=NULL,     The vector of colors to use for the lines. Default is rainbow()}

\item{draw.segments}{draw.segments=TRUE Whether or not horizontal segements should be drawn}

\item{draw.mid}{draw.mid=TRUE Whether or not a vertical line should be drawn down the middle of the profile plot.}

\item{plot.legend}{plot.legend=TRUE Whether or not to include a legend at the bottom}

\item{legend.n.col}{legend.n.col=4 The number of columns that should be used in the legend}

\item{sep}{sep=" - ""Aufregend - Langweilig"}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
Ein standart plot
}
\description{
This code is based on the the profile plotting code by Detlev Reymann October 2011
}
\examples{
## Create some random data
set.seed(0815)
x = matrix(rnorm(50), ncol=5)
colnames(x) = c("A", "B", "C", "D", "E")

## Some random labels for the left and right text. These labels
## need to be the same length as the number of rows in x
text.left = paste(1:nrow(x), "Left Text")
text.right = paste(1:nrow(x), "Right Text")

## Create the profile plot. See the details in profile.plot.R for
## the details on the parameters
labels<-c(
  "Aufregend - Langweilig",
  "Gut gemacht - Schlecht gemacht",
  "Glaubwürdig - Unglaubwürdig",
  "Spannend - Eintönig",
  "Verständlich - Unverständlich",
  "Guter Action-Schauspieler - schlechter Action Schauspieler",
  "sympathisch - unsympathisch",
  "talentiert - untalentiert",
  "modisch - unmodisch",
  "positive persönliche Einstellung - negative Einstellung gegenüber"
)



graphics.off()
rownames(x)<- labels
windows(8,8)
profile.plot(
  x,
  highlight.col=2,
  legend.n.col=5)

n<-10
data<- data.frame(Group=gl(2, n, labels = c("Control", "Treat")),
                  f1=rnorm(n*2),
                  f2=rnorm(n*2),
                  f3=rnorm(n*2),
                  f4=rnorm(n*2),
                  f5=rnorm(n*2),
                  f6=rnorm(n*2),
                  f7=rnorm(n*2),
                  f8=rnorm(n*2),
                  f9=rnorm(n*2),
                  f10=rnorm(n*2)
)

library(stp25)
x<-Recast2(.~Group, data, function(x) mean(x))
x$variable<- factor(x$variable, levels(x$variable), labels)



windows(8,8)
profile.plot( tidyr::spread(x, Group, value) ,
              draw.mid=FALSE
)
}
