% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{farbe}
\alias{standard_theme}
\alias{ggplot_theme}
\alias{bw_theme}
\alias{reset_lattice}
\alias{set_lattice}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stp25settings}{\code{\link[stp25settings:set_lattice]{bw_theme}}, \code{\link[stp25settings]{farbe}}, \code{\link[stp25settings:set_lattice]{ggplot_theme}}, \code{\link[stp25settings:set_lattice]{reset_lattice}}, \code{\link[stp25settings]{set_lattice}}, \code{\link[stp25settings:set_lattice]{standard_theme}}}
}}

