% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flow_chart.R
\name{simple_consort_plot}
\alias{simple_consort_plot}
\title{simple consort plot}
\usage{
simple_consort_plot(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{an  stp25plot:::visr}
}
\value{
plot
}
\description{
consort_plot
}
