% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_sentence.R
\name{wrap_sentence}
\alias{wrap_sentence}
\title{wrap_sentence}
\usage{
wrap_sentence(x, width = 25)
}
\arguments{
\item{x}{data.frame oder String}

\item{width}{default width= 25}
}
\value{
Character String
}
\description{
Kopie von \link{str_wrap} wobei die Labels mit
upData2 ergaenzt werden wenn ein Data.Frame-Objekt uebergeben wird
}
