% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_bwplot2.R
\name{bwplot2}
\alias{bwplot2}
\title{bwplot2}
\usage{
bwplot2(
  ...,
  xlab = NULL,
  ylab = NULL,
  auto.key = list(),
  box.width = 1/4,
  space_between = 1.2,
  par.settings = list(superpose.symbol = list(fill =
    lattice::trellis.par.get()$superpose.symbol$col)),
  panel = NULL
)
}
\arguments{
\item{...}{alles an lattice}

\item{box.width}{Box breite}

\item{space_between}{Box breite}

\item{panel}{eigene panel funktion zB mit  panel.abline(v=3)}

\item{nlevels}{anzahl an Gruppen}
}
\description{
Lattice bwplot mit groups. Ist eine erweiterung von lattice \link{bwplot}. Die Funktion arbeitet mit \link{panel.superpose}.
}
\examples{
set.seed(2)

res = data.frame(coef=rnorm(99) + (-1):1,
                 habitat=sample(letters[1:4], 99, TRUE),
                 grp=c("W", "X", "Y"))

bwplot2(
  coef ~ habitat,
  data = res,  groups = grp,
  main = "bwplot",
  pch = 15,
  space_between = 1.10,
  box.width = .15
)



}
