% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bwplot2.R
\name{bwplot2}
\alias{bwplot2}
\alias{bwplot2.data.frame}
\alias{bwplot2.formula}
\alias{panel.bwplot.groups}
\title{bwplot2}
\usage{
bwplot2(x, data, ...)

\method{bwplot2}{data.frame}(data, x, ...)

\method{bwplot2}{formula}(
  ...,
  xlab = NULL,
  ylab = NULL,
  auto.key = list(),
  box.width = 1/4,
  space_between = 1.2,
  par.settings = list(superpose.symbol = list(fill =
    lattice::trellis.par.get()$superpose.symbol$col)),
  panel = function(...) {
     dots <- list(...)
     lattice::panel.superpose(...,
    space_between = space_between, nlevels = nlevels(dots$groups))
 },
  panel.groups = panel.bwplot.groups
)

panel.bwplot.groups(x, y, ..., group.number, nlevels, space_between)
}
\arguments{
\item{...}{alles an lattice}

\item{box.width}{Box breite}

\item{space_between}{Box breite}

\item{panel}{eigene panel funktion zB mit  panel.abline(v=3)}

\item{nlevels}{anzahl an Gruppen}
}
\description{
Lattice bwplot mit groups. Ist eine erweiterung von lattice \link{bwplot}. Die Funktion arbeitet mit \link{panel.superpose}.
}
\details{
Die Funktion kann nicht mit den 
funktionen von lattice kombiniert werden.
}
\examples{
set.seed(2)

res = data.frame(coef=rnorm(99) + (-1):1,
                 habitat=sample(letters[1:4], 99, TRUE),
                 grp=c("W", "X", "Y"))

bwplot2(
  coef ~ habitat,
  data = res,  groups = grp,
  main = "bwplot",
  pch = 15,
  space_between = 1.10,
  box.width = .15
)



# bwplot2(
#   coef ~grp ,
#   data = res,  groups =habitat ,
#   main = "bwplot",
#   pch = 15,
#   space_between = 1.10,
#  box.width = .15,
#   auto.key = list(columns=3)
# )

bwplot2(
  habitat  ~ coef,
  data = res,
  groups = grp,
  main = "bwplot",
  pch = 15,
  auto.key = list(columns = 3),
  par.settings = stp25output::set_lattice_bw(col = grey.colors(4, start = 0.4, end = 0.9)),
  panel = function(...) {
    dots <- list(...)
    lattice::panel.abline(v = 0, lty = 1, col = "gray20")
    lattice::panel.superpose(...,
                             space_between = 1.1,
                             nlevels = nlevels(dots$groups))

    lattice::panel.abline(v = 0, lty = 3, col = "gray80")
  }
  )
}
