% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-cor_plot.R
\name{corr_plot}
\alias{corr_plot}
\alias{corr_plot.formula}
\alias{corr_plot.data.frame}
\alias{corr_pairs}
\title{Correlation Plot}
\usage{
corr_plot(x, ...)

\method{corr_plot}{formula}(x, data, cex.labels = NULL, ...)

\method{corr_plot}{data.frame}(
  x,
  ...,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  main = "",
  sub,
  xlab,
  ylab
)

corr_pairs(
  data,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  ...
)
}
\arguments{
\item{x}{formula oder data.frame}

\item{...}{an prepare_data2}

\item{data}{a data matrix}

\item{jitter}{Rauschen}

\item{smooth, lines}{Anpassungslienien}

\item{lines}{Regressinsgerade}

\item{pch}{Symbole  pch=20}

\item{digits, method}{correlation}

\item{cex.cor, resize}{Fixe groese mit cex.cor, resize abhaengig von r-Wert}

\item{method}{c("pearson", "kendall", "spearman")}

\item{stars, resize, cex.cor}{correlation formatierung}

\item{hist}{Histogram TRUE/FLASE}

\item{col.bar, col.bar.border, col.line, col.smooth}{Farben}

\item{smooth}{Gezeichnete Lineie}

\item{digits}{Nachkommastellen in plot}

\item{stars}{Sternchen}
}
\value{
nix
}
\description{
Function for making a correlation plot starting from a  formula and a data.frame
}
\examples{

#' 
require(stp25plot)
require(stp25tools)
require(tidyverse)

set.seed(1)

n <- 50

dat3 <- data.frame(a = rnorm(n)) \%>\%
  
  mutate(
    b = a +  rnorm(n),
    c = b +  rnorm(n),
    d = c +  rnorm(n),
    e = d +  rnorm(n),
    f = e + rnorm(n)
    
  )


corr_plot( ~ a + e + f+ b + c + d , dat3,
            resize=TRUE
)

#corr_plot( dat3,resize=TRUE)

corr_plot( dat3, a, b, c,
            resize=TRUE
)

}
