% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-cor_plot.R
\name{corr_plot}
\alias{corr_plot}
\alias{corr_plot.formula}
\alias{corr_plot.data.frame}
\alias{corr_pairs}
\alias{corr_plot2}
\title{Correlation Plot}
\usage{
corr_plot(x, ...)

\method{corr_plot}{formula}(x, data, cex.labels = NULL, ...)

\method{corr_plot}{data.frame}(
  x,
  ...,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  main = "",
  sub,
  xlab,
  ylab
)

corr_pairs(
  data,
  jitter = FALSE,
  smooth = FALSE,
  lines = TRUE,
  pch = 20,
  digits = 2,
  cex.cor = NULL,
  cex.labels = NULL,
  method = "pearson",
  stars = FALSE,
  resize = FALSE,
  hist = TRUE,
  col.bar = "RoyalBlue",
  col.bar.border = "lightblue",
  col.line = "blue",
  col.smooth = col.line,
  ...
)

corr_plot2(
  ...,
  main = "",
  method = "color",
  order = c("original", "AOE", "FPC", "hclust", "alphabet"),
  type = c("full", "lower", "upper"),
  diag = FALSE,
  sig.level = 0.2
)
}
\arguments{
\item{x}{formula oder data.frame}

\item{...}{an prepare_data2}

\item{data}{a data matrix}

\item{jitter}{Rauschen}

\item{smooth, lines}{Anpassungslienien}

\item{lines}{Regressinsgerade}

\item{pch}{Symbole  pch=20}

\item{digits, method}{correlation}

\item{cex.cor, resize}{Fixe groese mit cex.cor, resize abhaengig von r-Wert}

\item{method}{c("circle", "square", "ellipse", "number", "shade", "color", "pie"),}

\item{stars, resize, cex.cor}{correlation formatierung}

\item{hist}{Histogram TRUE/FLASE}

\item{col.bar, col.bar.border, col.line, col.smooth}{Farben}

\item{main}{titel}

\item{order}{c("original", "AOE", "FPC", "hclust", "alphabet"),}

\item{type}{c("full", "lower", "upper"),}

\item{diag}{FALSE,}

\item{sig.level}{signifikanz}

\item{smooth}{Gezeichnete Lineie}

\item{digits}{Nachkommastellen in plot}

\item{stars}{Sternchen}
}
\value{
nix
}
\description{
Function for making a correlation plot starting from a  formula and a data.frame
}
\examples{

require(stp25plot)
require(stp25tools)
require(tidyverse)
n <- 50

dat3 <- data.frame(a = rnorm(n)) |>
  mutate(
    b = a + rnorm(n),
    c = b + rnorm(n),
    d = c + rnorm(n),
    e = d + rnorm(n),
    f = e + rnorm(n)
  )

corr_plot( ~ a + e + f+ b + c + d, 
    dat3,
    resize = TRUE)

# example code
n <- 500
set.seed(1)

dat3 <- data.frame(a = rnorm(n)) |>
  mutate(
    b = a + rnorm(n),
    c = b / 2 + rnorm(n),
    d = c / 3 + rnorm(n),
    e = 2 - a + rnorm(n),
    fhshdeztrdfkjsk = e / 5  + rnorm(n)
  )
corr_plot2(dat3)
}
