% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_flowchart_plot.R
\name{flowchart_plot}
\alias{flowchart_plot}
\title{flowchart_plot}
\usage{
flowchart_plot(txt, txt_length = 30, cex = 0.9, arr.pos = 0.55,
  box.size = c(0.1, 0.15), ...)
}
\arguments{
\item{txt}{ist liste mit den Inhalten}

\item{txt_length}{Zeilenumbruch}

\item{cex}{schriftgroesse}

\item{arr.pos}{Pfeilspitze}

\item{box.size}{vektor der Box und zwar für links und rechts}

\item{...}{alles fuer reorder}
}
\description{
Uebernommen aus library(diagram)
}
\examples{
#library(stp25)
#'   windows(7,10)
txt<-"AAA" #LETTERS[sample.int(23,1)]
windows(8,8)
flowchart_plot(list(
  File=txt,
  Raw.x=c("B","C"),
  Clean.Up = c("D",paste("E",format(Sys.time(), "\%Y-\%m-\%d"))                        ),
  Analyse=c( "F", "G")

),
box.size=c(.1,.12) )
}
