% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{forest_plot}
\usage{
forest_plot(x, data = NULL, main = NULL, transform = TRUE, ...)
}
\arguments{
\item{x}{Model}

\item{data}{Daten}

\item{main}{Uebersrift}

\item{...}{alles an}
}
\value{
ggplot
}
\description{
Gestohlen von survminer::ggforest()
}
\examples{

# require(surv)
# require(survminer)
 require(survival)
 require(stpvers)

forest_plot(glm(status ~ sex + rx + adhere,
data = colon, family = binomial())) 
#forest_plot(coxph(Surv(time, status) ~ sex + rx + adhere,
#             data = colon), data=colon) 


}
