% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_table.R
\name{ggplot_table}
\alias{ggplot_table}
\title{ggplot_table}
\usage{
ggplot_table(
  gparam,
  gof = NULL,
  model = NULL,
  transform = TRUE,
  main = "",
  refLabel = "reference",
  use_label = TRUE,
  digits = 2,
  cpositions = c(0.02, 0.22, 0.4),
  fontsize = 0.7,
  col.text = 1,
  col.plot = "gray20"
)
}
\arguments{
\item{gparam}{Parameter Tabele mit var, estimate, p.value usw}

\item{gof}{Godnes of fit als Character}

\item{model}{model-Info}

\item{transform}{log-Transformation TRUE/FALSE}

\item{main}{Ueerschrift}

\item{refLabel}{Bezeichnung der Referenzklasse}

\item{use_label}{Label aus stpvers}

\item{digits}{Digits}

\item{cpositions}{Position}

\item{col.text, col.plot, fontsize}{an ggplot}
}
\value{
ggplot
}
\description{
In forest_plot( verwendet)
Gestohlen von survminer::ggforest()
}
