% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-fliegen-schiss-plot.R
\name{plotSigBars}
\alias{plotSigBars}
\alias{panel.sig.bars}
\title{Der Fliegen-Schiss-Plot}
\usage{
plotSigBars(
  x,
  include.stars = TRUE,
  shift = TRUE,
  cex.stars = 0.9,
  offset = 0.2,
  ...
)

panel.sig.bars(
  x,
  include.stars = TRUE,
  shift = TRUE,
  cex.stars = 0.9,
  offset = 0.4,
  ...
)
}
\arguments{
\item{x}{Objekt lm oder emmeans}

\item{shift}{oben unden oder links rechts}

\item{cex.stars}{groesse}

\item{...}{nicht benutzt}

\item{stars}{Sternchen oder P-Weret}
}
\value{
plot-text
}
\description{
Der Fliegen-Schiss-Plot
}
\examples{

# kommt noch


require(lattice)
require(latticeExtra)
require(effects)


stripplot(
DC4 ~ time2,
data = dat, ylim=c(-10,70),
ylab = "Difference (Absolute Value) [mm]", jitter.data=TRUE,
panel = function(x, y, ...) {
  # panel.conf.int(x, y, ...)
  panel.stripplot(x,y, pch=19, col="gray50",...)
  #panel.points(x,y, pch=19,...)
  #panel.mean(x,y,  ...)
  
  panel.median(x,y, ...)
  panel.sig.bars(fit1, include.stars = FALSE, offset = .4)
}
)



fit1 <- lm(DC4 ~ time2, data = dat)

p <- stripplot(  DC4 ~ time2,  data = dat, jitter.data=TRUE,pch=20, col="gray50")
p + layer(panel.sig.bars(fit1, include.stars = FALSE, offset = .4) )

p2<- plot(effect("time2", fit1), ylim=c(0,60))
p2 +  layer(panel.sig.bars(fit1, include.stars = FALSE, offset = .4) )

#' 
#' 
}
